All Substrings of a String
A substring in Java is a sequence of characters extracted from a larger string.
In this code we are taking a user input string and finding all the substrings of it by defining a function named allsubstrings. In the function we are using 2 for loops and an inbuilt substring function to print all the substrings. And at the end we are also printing the count of number of substrings of the string.